Refactor document input
