+++ src/components/RouteGuard/RouteGuard.tsx
@@ -34,7 +34,7 @@ const RouteGuard: FC<RouteGuardProps> = ({ children }) => {
       "/producers",
       "/producers/:id",
       "/edit-production-company",
-      "/register-production-company",
+      "/producers/register",
       "/gardel-awards",
       "/audit-changes",
       "/audit-sessions",
@@ -73,7 +73,7 @@ const RouteGuard: FC<RouteGuardProps> = ({ children }) => {
       "/users/add-user",
       "/producers",
       "/producers/:id",
-      "/register-production-company",
+      "/producers/register",
       "/edit-production-company",
       "/gardel-awards",
       "/audit-changes",
@@ -106,7 +106,7 @@ const RouteGuard: FC<RouteGuardProps> = ({ children }) => {
       "/users",
       "/users/add-user",
       "/cashflow-account-statement",
-      "/register-production-company",
+      "/producers/register",
       "/my-profile",
       "/change-password",
     ],
@@ -119,7 +119,7 @@ const RouteGuard: FC<RouteGuardProps> = ({ children }) => {
       "/conflicts",
       "/users",
       "/cashflow-account-statement",
-      "/register-production-company",
+      "/producers/register",
       "/my-profile",
       "/change-password",
     ],
